pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
names = "^0.3.0"
djangorestframework = "^3.12.4"
[tool.poetry.dev-dependencies]
